Maven编译时发现这几个类报错，做等价修改

+++ src/com/funs/login/dao/LoginDAO.java
@@ -25,7 +25,7 @@ public class LoginDAO extends BaseDAO{
 	}
 	
 	public int queryIdForPackageVO(PackageVO packageVO) {
-		return this.sqlSessionTemplate.selectOne("com.funs.package.queryIdForPackageVO", packageVO);
+		return (Integer)this.sqlSessionTemplate.selectOne("com.funs.package.queryIdForPackageVO", packageVO);
 	}
 	
 	public void insertPackageReShop(PackageVO packageVO){
